Extract abstract class for quiz progress
