+++ includes/internal/student-progress/lesson-progress/models/class-tables-based-lesson-progress.php
@@ -7,8 +7,6 @@
 
 namespace Sensei\Internal\Student_Progress\Lesson_Progress\Models;
 
-use DateTimeInterface;
-
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
